+++ src/main/java/com/driver/services/UserService.java
@@ -2,6 +2,7 @@ package com.driver.services;
 
 import com.driver.models.*;
 import com.driver.repositories.BlogRepository;
+import com.driver.repositories.ImageRepository;
 import com.driver.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,13 +15,13 @@ import java.util.Optional;
 public class UserService {
     @Autowired
     private UserRepository userRepository3;
-
+    @Autowired
+    private ImageRepository imageRepository;
     @Autowired
     private BlogRepository blogRepository;
     public User createUser(String username, String password){
-            User newUser=new User();
-            newUser.setUsername(username);
-            newUser.setPassword(password);
+            User newUser=new User(username,password);
+
             userRepository3.save(newUser);
 
     return newUser;
@@ -29,17 +30,26 @@ public class UserService {
     public void deleteUser(int userId){
         //delete user and respective blog
         Optional<User> uu=userRepository3.findById(userId);
+        if(uu.isEmpty()){
+            return ;
+        }
         User user=uu.get();
 
         List<Blog> temp= user.getBlogList();
 
         userRepository3.deleteById(userId);
 
-        for(Blog i:temp){
+       /* for(Blog i:temp){
+            List<Image> ll=i.getImageList();
+
+            for(Image j:ll){
+                imageRepository.deleteById(j.getId());
+            }
             blogRepository.deleteById(i.getId());
 
 
-        }
+
+        }*/
 
 
     }
@@ -48,8 +58,10 @@ public class UserService {
 
             Optional<User> user=userRepository3.findById(id);
 
+
             User uu=user.get();
             uu.setPassword(password);
+            userRepository3.save(uu);
 
             return uu;
